Rework DNS hosts to enable the use of 'count'
This creates a new datasource called libvirt_network_dns_hosts_template,
which can be used to populate the dns_host attribute in libvirt_network
resources. The use of a datasource is required in order to dynamically
generate dns_host entries (since 'count' is the only looping construct
in Terraform and 'count' is limited to datasources and resources). The
list of hostnames has also been replaced by a single hostname with the
option of declaring multiple dns_host blocks with the same IP address.

I originally tried to externally generate a variable that would host the
list of maps needed for the dns_host attribute, but Terraform explicitly
disallows lists of maps greater than length one. Instead, I decided to
follow the pattern established by template_file and create a resource
that can be rendered.

The hostnames attribute had to be replaced with hostname because the
underlying schema for the 'rendered' attribute is a TypeMap. In
Terraform, this is not a generic map, but is instead limited to
map[string]string. To accommodate this, multiple dns_host blocks can be
defined with the same IP address, one for each hostname.